re-organize test files so that their name corresponds to the file they are testing
