+++ frontend/src/pages/modelServing/screens/global/__tests__/utils.spec.ts
